requirements.txt
# Anthropic SDK + Agent SDK
anthropic>=0.40.0

# MCP server framework
fastmcp>=0.4.0

# MCP client (stdio transport used by orchestrator)
mcp>=1.0.0

# Data processing
numpy>=1.26.0
pandas>=2.1.0

# Environment variable loading
python-dotenv>=1.0.0

# File watching (auto-ingest on new .ld file)
watchdog>=4.0.0

# Testing
pytest>=8.0.0
